backend/news/requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
anthropic
pydantic>=2
python-dotenv